+++ src/index.ts
@@ -16,3 +16,4 @@ export * from './wrapper';
 export * from './either';
 export * from './maybe';
 export * from './list';
+export * from './reviver';